Change login and registration form (#139)
* change login and registration form

* Update themeOrder.js

* Update themeOrder.js
