+++ client/src/App.jsx
@@ -6,6 +6,8 @@ import { PersistGate } from 'redux-persist/integration/react';
 import Basket from './pages/Basket';
 import HomeContent from './pages/Home';
 import Header from './components/Header';
+import Login from './pages/Login';
+import Registration from './pages/Registration';
 import ProductDescription from './components/ProductDescription';
 import PageNotFound from './components/NotFoundPage';
 import store, { persistor } from './redux/store';
@@ -27,6 +29,8 @@ function App() {
                 <Route path="/about" element={<About />} />
                 <Route path="/user" element={<UserPage />} />
                 <Route index element={<HomeContent />} />
+                <Route path="/registration" element={<Registration />} />
+                <Route path="/login" element={<Login />} />
                 <Route path="/basket" element={<Basket />} />
                 <Route path="/product" element={<ProductsContent />} />
                 <Route path="/product/:id" element={<ProductDescription />} />
