+++ client/src/themes/theme.js
@@ -16,6 +16,7 @@ const theme = createTheme({
     secondary: {
       main: '#ff6565',
       contrastText: '#f5f7fb',
+      inputBackground: '#F3F6FF',
       green: '#1b7a0f',
       greenTransparent: '#1b7a0f47',
     },
